Swapped CSS Filters for jQuery
The CSS filters were super cool, but a bit over the top and not
terribly functional.  The navigation to the anchor caused all sorts
of weird jumping and then caused extra compensation for that.  Showing
and hiding objects with JS is cleaner and ultimately makes the page
smaller to load.  It's a better, but less cool solution.