Moved About Me to it's own page
The about me contents of the home page were moved into
their own page.  This includes the bio section, professional history,
and the contact me section.  All of these sections had some changes
and styling to update the rest of the site.

The milestone sections were updated to have similar text styling as
the blog posts (albiet smaller font) and the biography section
has the same styling as the blog posts.

On the home page a list of recent blog posts was added (currently all)
and the main landing area was updated to include typed.js so that the
descriptive text will change automatically.

The layouts were updated and a frame layout was created.  This small
change allows for the About Me page to have the same banner logic as
Blog and Projects, but also allows the main content section to be
different. This split makes the full width sections of the About Me
page possible.